% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svgviewr.R
\name{svgviewr}
\alias{svgviewr}
\title{<Add Title>}
\usage{
svgviewr(svg = "", animation_reverse = FALSE, animation_duration = NULL,
  animation_repeat = NULL, animation_count = NULL, margin = list(),
  background_color = NULL, start_rotate = FALSE,
  show_control_panel = TRUE, width = NULL, height = NULL,
  elementId = NULL)
}
\description{
<Add Description>
}

