\name{svgViewR-package}
\alias{svgViewR}

\alias{apply_transform_svg}
\alias{applyTransform_svg}
\alias{applyTransformations}
\alias{applyQuat}
\alias{axisAngle2Quat}
\alias{axisAngle2Tmat}
\alias{avec_svg}
\alias{bestAlign_svg}
\alias{box_in_box_faces}
\alias{cameraParameters}
\alias{circlePoint_svg}
\alias{cprod_svg}
\alias{create_plane_mesh}
\alias{create_sphere_mesh}
\alias{darkenCol}
\alias{default_gpar}
\alias{defineCircle_svg}
\alias{distPointToPlane_svg}
\alias{distancePointToLine_svg}
\alias{dppt_svg}
\alias{findPlotDims}
\alias{fitShapes}
\alias{html2plot}
\alias{image2world}
\alias{JSONToobj}
\alias{lightenCol}
\alias{lim2corners}
\alias{mtransform_svg}
\alias{plot_svg_shapes}
\alias{pointLineProj_svg}
\alias{pointPlaneProj_svg}
\alias{pointNormalOnLine_svg}
\alias{quat2RM}
\alias{Rhttpd2}
\alias{Rhttpd2-class}
\alias{read_obj_str}
\alias{readCam}
\alias{readHTML}
\alias{RM2Quat}
\alias{rm2euler}
\alias{sd_linkr}
\alias{splitAlphaNum}
\alias{svgviewr.write}
\alias{svg.arrow}
\alias{svg.box}
\alias{svg.camera}
\alias{svg.carrow}
\alias{svg.ccuboid}
\alias{svg.circles}
\alias{svg.cone}
\alias{svg.coupler}
\alias{svg.cplanes}
\alias{svg.cuboid}
\alias{svg.cylinder}
\alias{svg.dcylinder}
\alias{svg.deform}
\alias{svg.farrow}
\alias{svg.frame}
\alias{svg.hcuboid}
\alias{svg.images}
\alias{svg.open}
\alias{svg.pal}
\alias{svg.paths}
\alias{svg.plane}
\alias{svg.pointsC}
\alias{svg.socket}
\alias{svg.sphere}
\alias{svg.spheres}
\alias{svg.text}
\alias{svg.triangle}
\alias{svg.rotate}
\alias{svg.translate}
\alias{svg.transform}
\alias{svg_axis_grids}
\alias{svg_axis_polygons}
\alias{svg_axis_ticks}
\alias{svg_box_lim}
\alias{svg_ranges}
\alias{svgviewr.circles}
\alias{svgviewr.lines}
\alias{svgviewr.new}
\alias{svgviewr.paths}
\alias{svgviewr.pathsC}
\alias{svgviewr.points}
\alias{svgviewr.pointsC}
\alias{svgviewr.text}
\alias{svgviewr_env}
\alias{svgviewr_ranges}
\alias{tMatrixEP_svg}
\alias{tm2JSON}
\alias{uvector_svg}
\alias{viewer_save_image}
\alias{vorthogonal_svg}
\alias{webColor}
\alias{write_HTML}
\alias{writeOBJ}

\docType{package}
\title{3D Animated Interactive Visualizations using SVG}
\description{The \code{svgViewR} package allows users to create 3D animated, interactive visualizations that can be viewed in a web browser. Please see \href{https://aaronolsen.github.io/tutorials/visualization3d.html}{svgViewR tutorials} for examples of how to use this package.}
\details{
\tabular{ll}{
Package: \tab svgViewR\cr
Type: \tab Package\cr
Version: \tab 1.4\cr
Date: \tab 2019-07-03\cr
License: \tab GPL (>= 2)\cr}}
\author{
Aaron Olsen (maintainer, <aarolsen@gmail.com>), Alejandro Romero
}