\name{svgViewR-package}
\alias{svgViewR}

\alias{applyTransform_svg}
\alias{applyTransformations}
\alias{apply_transform_svg}
\alias{cprod_svg}
\alias{create_sphere_mesh}
\alias{default_gpar}
\alias{findPlotDims}
\alias{fitShapes}
\alias{html2plot}
\alias{lim2corners}
\alias{mtransform_svg}
\alias{plot_svg_shapes}
\alias{Rhttpd2}
\alias{Rhttpd2-class}
\alias{readHTML}
\alias{rm2euler}
\alias{sd_linkr}
\alias{splitAlphaNum}
\alias{svgviewr.write}
\alias{svg.box}
\alias{svg.circles}
\alias{svg.coupler}
\alias{svg.cuboid}
\alias{svg.cylinder}
\alias{svg.frame}
\alias{svg.open}
\alias{svg.paths}
\alias{svg.pointsC}
\alias{svg.socket}
\alias{svg.sphere}
\alias{svg.spheres}
\alias{svg.text}
\alias{svg.rotate}
\alias{svg.translate}
\alias{svg.transform}
\alias{svg_axis_grids}
\alias{svg_axis_polygons}
\alias{svg_axis_ticks}
\alias{svg_box_lim}
\alias{svg_ranges}
\alias{svgviewr.circles}
\alias{svgviewr.lines}
\alias{svgviewr.new}
\alias{svgviewr.paths}
\alias{svgviewr.pathsC}
\alias{svgviewr.points}
\alias{svgviewr.pointsC}
\alias{svgviewr.text}
\alias{svgviewr_env}
\alias{svgviewr_ranges}
\alias{tMatrixEP_svg}
\alias{tm2JSON}
\alias{uvector_svg}
\alias{vorthogonal_svg}
\alias{webColor}
\alias{write_HTML}

\docType{package}
\title{3D Animated Interactive Visualizations using SVG}
\description{The \code{svgViewR} package allows users to create 3D animated, interactive visualizations that can be viewed in a web browser.}
\details{
\tabular{ll}{
Package: \tab svgViewR\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2018-01-19\cr
License: \tab GPL (>= 2)\cr}}
\author{
Aaron Olsen
Maintainer: Aaron Olsen <aarolsen@gmail.com>
}