\name{svgViewR-package}
\alias{svgviewr.write}

\docType{package}
\title{3D Animated Interactive Visualizations using SVG}
\description{The \code{svgViewR} package allows users to write animated, 2D and 3D shapes into an SVG Viewer HTML file using functions similar to the native R plotting functions. The SVG Viewer file can be opened with any compatible web browser (Safari, Firefox, Chrome, Opera) and provides interactive zoom, 3D rotation and translation using mouse and keyboard shortcuts (see \code{\link{svgviewr.new}}). This package currently supports points, lines and simple paths. Future versions will support additional shapes.}
\details{
\tabular{ll}{
Package: \tab svgViewR\cr
Type: \tab Package\cr
Version: \tab 1.0.1\cr
Date: \tab 2014-12-27\cr
License: \tab GPL (>= 2)\cr}}
\author{
Aaron Olsen
Maintainer: Aaron Olsen <aolsen@uchicago.edu>
}